Add files via upload
This is a project I created as a college project. The basic idea was to create an algorithm to help users find a perfect match for their dogs as mating is important for your pet. Currently this project has two features namely, matching algorithm and a personalised blog section which I created from scratch. I have a lot of great ideas about the features that I wish to add to this project using different frameworks and tools as I learn more about them. I have kept this project open so that you can guide me in making any changes if you see any room for improvement. I am still learning my way around GitHub and I hope this turns out great. :)
As this is my first project, I am open to suggestions and comments.