Add files via upload
This is Playdate. 

Playdate helps you find a suitable mate for your pet dog. Finding a mate for your dog that is suitable is actually a problem that, if not taken care of can lead to the birth of unhealthy puppies. These problems can be taken care of by considering a few parameters that I have included in my main matching algorithm.

Another part of this upload is the personalised blog section which has data for one entry as of now but I will be adding more and more as I collect more data.

This project right now is written in Python but I want to learn and implement Django as a framework so that I can ultimately create a working web application that can be used in every dog's household. :)

I see potential in this project to expand and have various features like:

1. Covid Checker - Checking the number of covid cases in your state so you can see if it is safe to travel.
2. Security Layer - I intend on experimenting with different kinds of security systems. Possibly voice recognition, learning how to integrate various image processing libraries and some of the basic APIs such as Catchpa for safety.
3. More Pets - This project has the potential to be helpful for different animals as well.
4. Adoption - Removing the middlemen such as brokers who mistreat and often are abusive towards the puppies or baby animals, the user can put the puppies directly up for adoption and can directly reach families that actually want to care for the innocent animals.
5. Community Help - A community section can be added to the web application where there will be suggestions from actual verified pet owners and a place for influencers where they can bring pet owners closer to local businesses and brands to promote the Vocal for Local movement.

This is my first project and I hope to learn more and more and add more components to this project that feel relevant to the project. I want to hopefully publish this project one day and help a great number of families and our loyal friends of course. Please feel free to share knowledge with me which can be the different technologies I can use to accomplish what I aspire for this web app to become.